Working airspy spectrum but audio distorted

+++ src/decoder/FmDemodulator.java
@@ -37,12 +37,12 @@ import common.Config;
  *
  */
 public class FmDemodulator {
-	protected double[] i = new double[3];
-	protected double[] q = new double[3];
-	double lastPhase = 0d;
-	double gain = 0.5;
-	double limiti;
-	double limitq;
+	protected float[] i = new float[3];
+	protected float[] q = new float[3];
+	float lastPhase = 0f;
+	float gain = 0.5f;
+	float limiti;
+	float limitq;
 	
 	
 	public FmDemodulator() {
@@ -67,9 +67,9 @@ public class FmDemodulator {
 	 * @param qn
 	 * @return
 	 */
-	public double atanDemodulate(double in, double qn) {
-		double I,Q; 
-		double angle = 0;;
+	public double atanDemodulate(float in, float qn) {
+		float I,Q; 
+		float angle = 0;;
 		
 		/**
 		 * Multiply the current sample against the complex conjugate of the 
@@ -83,7 +83,7 @@ public class FmDemodulator {
 		q[0] = qn;
 		
 		if (I == 0) {
-			I=1E-20; // small value to prevent divide by zero error
+			I=1E-20f; // small value to prevent divide by zero error
 			//angle = 0;
 		} 
 		/**
@@ -92,15 +92,15 @@ public class FmDemodulator {
 		 * original message waveform during the modulation.  This value now
 		 * serves as the instantaneous amplitude of the demodulated signal
 		 */
-		double denominator = 1.0d / I;
-		angle = Math.atan( (double)Q * denominator );
+		float denominator = 1.0f / I;
+		angle = (float) Math.atan( (float)Q * denominator );
 
 		// If both real and imaginary parts are negative, need to subtract PI radians
 		if (I < 0 && Q < 0) {
-			angle = angle - Math.PI;
+			angle = (float) (angle - Math.PI);
 		}
 		if (I < 0 && Q >= 0) {
-			angle = angle + Math.PI;
+			angle = (float) (angle + Math.PI);
 		}
 
 		return angle * gain;
@@ -143,7 +143,7 @@ public class FmDemodulator {
 	 * @param q
 	 * @return
 	 */
-	public double demodulate(double in, double qn) {
+	public float demodulate(float in, float qn) {
 		if (Config.useLimiter) {
 		limiter(in, qn);
 		
@@ -163,12 +163,12 @@ public class FmDemodulator {
 		
 		// it simplifies to: Demodn={Qn*In-1  -  In*Qn-1}/{In2+Qn2}
 		
-		double gain = 0.5d; // magic number of 1/2 seems to work best
-		double num = i[1] * ( q[2] - q[0] ) - q[1] * ( i[2] - i[0] );
+		float gain = 0.5f; // magic number of 1/2 seems to work best
+		float num = i[1] * ( q[2] - q[0] ) - q[1] * ( i[2] - i[0] );
 		//double num = q[2] * i[1]  - i[2] * q[1];
-		double den = (i[1]*i[1] + q[1]*q[1]); 
+		float den = (i[1]*i[1] + q[1]*q[1]); 
 		
-		double deltafreq =  gain* (num/den);
+		float deltafreq =  gain* (num/den);
 	
 		if (Double.isNaN(deltafreq)) {
 			//Log.println("FM NaN: num:"+num+" den:"+den);
@@ -182,10 +182,10 @@ public class FmDemodulator {
 	 * @param i
 	 * @param q
 	 */
-	private void limiter(double i, double q) {
-		double f = Math.atan2(q, i);
-		limiti = Math.cos(f);
-		limitq = Math.sin(f);
+	private void limiter(float i, float q) {
+		float f = (float) Math.atan2(q, i);
+		limiti = (float) Math.cos(f);
+		limitq = (float) Math.sin(f);
 	}
 
 }
